+++ src/os/common/CriticalSection.cpp
@@ -1,0 +1,22 @@
+// Copyright (C) 2023 Igor Krechetov
+// Distributed under MIT license. See file LICENSE for details
+#include "hsmcpp/os/common/CriticalSection.hpp"
+
+#include "hsmcpp/os/InterruptsFreeSection.hpp"
+#include "hsmcpp/os/Mutex.hpp"
+
+namespace hsmcpp {
+
+CriticalSection::CriticalSection(Mutex& sync)
+    : mSync(sync) {
+    mInterruptsBlock = new InterruptsFreeSection();
+    mSync.lock();
+}
+
+CriticalSection::~CriticalSection() {
+    mSync.unlock();
+    delete mInterruptsBlock;
+    mInterruptsBlock = nullptr;
+}
+
+}  // namespace hsmcpp